+++ CHANGELOG.rst
@@ -43,6 +43,7 @@ All notable changes to this project are documented in this file.
 - Fix ``BigInteger.ToByteArray()`` for some negative values to return too many bytes
 - Implement SimplePolicyPlugin for transactions sent to a node `#960 <https://github.com/CityOfZion/neo-python/issues/960>`_
 - Fix transaction deserialization not setting correct type for ``ContractTransaction``
+- Fix ``GetBigInteger()`` return value of ``Boolean`` ``StackItem`` to return correct type
 
 
 [0.8.4] 2019-02-14